backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
python-multipart
PyPDF2
PyMuPDF
pdfplumber
